Move file cache to separate package
